+++ prefect_azure/__init__.py
@@ -3,6 +3,8 @@ from .credentials import (  # noqa
     AzureBlobStorageCredentials,
     AzureCosmosDbCredentials,
     AzureMlCredentials,
+    AzureKeyVaultCredentials,
+    AzureKeyVaultSecretReference,
 )
 
 __version__ = _version.get_versions()["version"]